[ref] deprecated stupid dynamic type for slider, now type-specification based.

+++ include/ekg/util/text.hpp
@@ -78,12 +78,12 @@ namespace ekg {
   /**
    * Returns a customised float64 precision as a string.
    **/
-  std::string_view string_float64_precision(double number, uint8_t precision);
+  std::string_view string_float64_precision(double number, int32_t precision);
 
   /**
    * Returns a customised float precision as a string.
    **/
-  std::string_view string_float_precision(float number, uint8_t precision);
+  std::string_view string_float_precision(float number, int32_t precision);
 
   /**
    * Return the index of last attribute-token and insert bits.
